Test watcher
A hacky but working implementation of the test runner. This
reruns all tests after a re-compile, firing on the callback
webpack gives us for that. For the time, re-running all tests
shouldn't be a problem since we don't have that many tests,
but I can imagine this getting noisy fast.