Sentry release uploads
Feature:
This adds Sentry uploads to our build step. These uploads are compiled
js files and their sourcemaps. Note, this doesn't do any of the Sentry
configuration or error handling for the project. This simply uploads the
needed files for readable stacktraces ahead of the actual exception.

How this works:
1. Sentry has a notion of "releases", which is just a versioned
collection of the compiled js files and their maps. This patch makes no
assumption on release versioning/naming and allows the projects using
`node-build` to choose their method.
2. We tell Sentry to create a new release with the given version.
3. We then upload our files to that release.
4. We add the release version to webpack's config for future
substitution into the project's codebase, where it will be used for
configuration.

Additional Details:
- Right now this only works for production client code. This leaves out
  backend and dev code. We can add backend in a future patch.
- This adds superagent as a dep in node-build.
- This requires two new env variables, `SENTRY_RELEASE_ENDPOINT` and
  `SENTRY_KEY`, to be set. This is the endpoint for creating releases
  and the auth token for that backend respectively.
- I've made our default webpack config files export functions instead of
  objects. This was done to make it straightforward to add the project
  name and release version to the config but it has the added benefit of
  making other future changes simpler. This was a conscious decision to
  not use a merging type strategy b/c merging is cumbersome and feels
  less obvious. Other than adding those two things to ProductionClient
  and renaming these files to more of a function-type name, nothing else
  has changed.