Sentry release uploads
Feature:
This adds Sentry uploads to our build step. These uploads are compiled
js files and their sourcemaps. Note, this doesn't do any of the Sentry
configuration or error handling for the project. This simply uploads the
needed files for readable stacktraces ahead of the actual exception.

How this works:
1. Sentry has a notion of "releases", which is just a versioned
collection of the compiled js files and their maps. This patch makes no
assumption on release versioning/naming and allows the projects using
`node-build` to choose their method.
2. We tell Sentry to create a new release with the given version.
3. We then upload our files to that release.
4. We add the release version to webpack's config for future
substitution into the project's codebase, where it will be used for
configuration.

Additional Details:
- Source maps are not supported in Node so we don't worry about the
  server.
- This adds superagent as a dep in node-build.
- This requires two new env variables, `SENTRY_RELEASE_BASE_URL` and
  `SENTRY_KEY`, to be set. This is the endpoint for creating releases
  and the auth token for that backend respectively.